requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<4
djangorestframework-simplejwt>=5.3,<6
django-environ>=0.11,<0.12
django-cors-headers>=4.4,<5
httpx>=0.27,<1
cryptography>=42,<43
python-dotenv>=1,<2
pytest>=7,<9
black>=23,<26
isort>=5,<7
flake8>=6,<8
flake8-docstrings>=1,<2
pep8-naming>=0.13,<0.14
mypy>=1,<2
